src/backend/requirements.txt
flask
tensorflow
pillow
numpy
pandas
rapidfuzz
prometheus-client
gunicorn
